refactor(tests): remove obsolete test modules and consolidate test cases into dedicated files for improved organization and maintainability
