+++ crates/dbx-api/src/tests/auth_tests/mod.rs
@@ -1,0 +1,10 @@
+//! Authentication and authorization test modules
+
+/// API key management tests
+mod api_key_tests;
+
+/// Permission system tests
+mod permission_tests;
+
+/// Role-based access control tests
+mod rbac_tests;
