+++ crates/dbx-api/src/tests/middleware_tests/mod.rs
@@ -1,0 +1,7 @@
+//! Middleware test modules
+
+/// Rate limiting middleware tests
+mod rate_limit_tests;
+
+/// Security middleware tests
+mod security_tests;
