+++ crates/dbx-api/src/tests/mod.rs
@@ -8,3 +8,21 @@ pub mod utils;
 
 /// Model serialization and validation tests  
 mod model_tests;
+
+/// Configuration and environment tests
+mod config_tests;
+
+/// Authentication and authorization tests
+mod auth_tests;
+
+/// Route handler tests
+mod route_tests;
+
+/// Server initialization and lifecycle tests
+mod server_tests;
+
+/// Middleware functionality tests
+mod middleware_tests;
+
+/// Main application tests
+mod main_tests;
