+++ crates/dbx-api/src/tests/route_tests/mod.rs
@@ -1,0 +1,7 @@
+//! Route handler test modules
+
+/// API key route tests
+mod api_key_route_tests;
+
+/// Role management route tests
+mod role_route_tests;
